requirements.txt
fastapi
uvicorn
# Haystack 2.0 and components
haystack-ai
weaviate-haystack
sentence-transformers
# Utils
gitpython
pydantic-settings
pyyaml
python-multipart
# Testing
pytest
httpx
# Tracing
arize-phoenix
openinference-instrumentation-haystack
ollama-haystack
google-ai-haystack
accelerate

# Tree-sitter
tree-sitter
tree-sitter-python
tree-sitter-javascript
tree-sitter-typescript
tree-sitter-java
tree-sitter-c-sharp
tree-sitter-php
tree-sitter-go

-e .
